Refactor RemoveAssociationSweeper into one template class for use in both DataReaderImpl and RecorderImpl
