+++ dds/DCPS/DataReaderImpl.h
@@ -38,6 +38,7 @@
 #include "ReactorInterceptor.h"
 #include "Service_Participant.h"
 #include "PoolAllocator.h"
+#include "RemoveAssociationSweeper.h"
 
 #include "ace/String_Base.h"
 #include "ace/Reverse_Lock_T.h"
@@ -112,61 +113,6 @@ public:
 };
 
 #endif
-
-// Class to cleanup associations scheduled for removal
-class RemoveAssociationSweeper : public ReactorInterceptor {
-public:
-  RemoveAssociationSweeper(ACE_Reactor* reactor,
-                                  ACE_thread_t owner,
-                                  DataReaderImpl* reader);
-
-  void schedule_timer(OpenDDS::DCPS::RcHandle<OpenDDS::DCPS::WriterInfo>& info, bool callback);
-  void cancel_timer(OpenDDS::DCPS::RcHandle<OpenDDS::DCPS::WriterInfo>& info);
-
-  // Arg will be PublicationId
-  int handle_timeout(const ACE_Time_Value& current_time, const void* arg);
-
-  virtual bool reactor_is_shut_down() const
-  {
-    return TheServiceParticipant->is_shut_down();
-  }
-
-private:
-  ~RemoveAssociationSweeper();
-
-  DataReaderImpl* reader_;
-
-  class CommandBase : public Command {
-  public:
-    CommandBase(RemoveAssociationSweeper* sweeper,
-                OpenDDS::DCPS::RcHandle<OpenDDS::DCPS::WriterInfo>& info)
-      : sweeper_ (sweeper)
-      , info_(info)
-    { }
-
-  protected:
-    RemoveAssociationSweeper* sweeper_;
-    OpenDDS::DCPS::RcHandle<OpenDDS::DCPS::WriterInfo> info_;
-  };
-
-  class ScheduleCommand : public CommandBase {
-  public:
-    ScheduleCommand(RemoveAssociationSweeper* sweeper,
-                    OpenDDS::DCPS::RcHandle<OpenDDS::DCPS::WriterInfo>& info)
-      : CommandBase(sweeper, info)
-    { }
-    virtual void execute();
-  };
-
-  class CancelCommand : public CommandBase {
-  public:
-    CancelCommand(RemoveAssociationSweeper* sweeper,
-                  OpenDDS::DCPS::RcHandle<OpenDDS::DCPS::WriterInfo>& info)
-      : CommandBase(sweeper, info)
-    { }
-    virtual void execute();
-  };
-};
 
 // Class to cleanup in case EndHistoricSamples is missed
 class EndHistoricSamplesMissedSweeper : public ReactorInterceptor {
@@ -727,7 +673,7 @@ private:
 
   friend class InstanceState;
   friend class EndHistoricSamplesMissedSweeper;
-  friend class RemoveAssociationSweeper;
+  friend class RemoveAssociationSweeper<DataReaderImpl>;
 
   friend class ::DDS_TEST; //allows tests to get at private data
 
@@ -738,7 +684,7 @@ private:
   SubscriberImpl*              subscriber_servant_;
   DDS::DataReader_var          dr_local_objref_;
   EndHistoricSamplesMissedSweeper* end_historic_sweeper_;
-  RemoveAssociationSweeper*        remove_association_sweeper_;
+  RemoveAssociationSweeper<DataReaderImpl>*        remove_association_sweeper_;
 
   CORBA::Long                  depth_;
   size_t                       n_chunks_;
